No case-specific printing styles ♻️
Good printing styles merged with the main files,
i.e. styles for links placed into `link.css`.

Media toggles are refactored and currently considered experimental.
Those are not used almost anywhere and seem to be not needed.
There is a `media/toggle.css` that imposedl quite tech API
and is not included into the default style under `index.css`.